requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
stix2==3.0.1
stix2-patterns==2.0.0
urllib3==2.0.2
